refactor: move event types to type folder
