+++ packages/aurora/src/types/Events.ts
@@ -1,5 +1,5 @@
-import { Toast, ToastWithoutIdAndType } from '../types/Toast'
-import { ToastId } from '../types/ToastId'
+import type { Toast, ToastWithoutIdAndType } from './Toast'
+import { type ToastId } from './ToastId'
 
 export interface CustomEventDetail extends Event {
   detail: Toast
